rename models and add product entity and dtos
