+++ src/main/java/com/bhaskar/store/management/services/impl/UserServiceImpl.java
@@ -4,7 +4,7 @@ import com.bhaskar.store.management.controllers.UserController;
 import com.bhaskar.store.management.dtos.PageableResponse;
 import com.bhaskar.store.management.dtos.UserDto;
 import com.bhaskar.store.management.exceptions.ResourceNotFoundException;
-import com.bhaskar.store.management.models.User;
+import com.bhaskar.store.management.entity.User;
 import com.bhaskar.store.management.repositories.UserRepo;
 import com.bhaskar.store.management.services.UserService;
 import com.bhaskar.store.management.utility.Util;
@@ -25,7 +25,6 @@ import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
